+++ tests/testthat/test-missing.R
@@ -14,8 +14,8 @@ test_that("original data with NAs do not pose problems in glm and lm.", {
 test_that("newdata with NAs do not pose problems in lm.", {
     mod <- lm(hp ~ mpg + drat + wt + factor(gear), data = tmp)
 
-typical(model = mod, drat = c(NA, 10))
+datagrid(model = mod, drat = c(NA, 10))
 
-    mfx <- marginaleffects(mod, newdata = typical(drat = c(NA, 10)))
+    mfx <- marginaleffects(mod, newdata = datagrid(drat = c(NA, 10)))
     expect_s3_class(tidy(mfx), "data.frame")
 })